feat: big change from 32 to 64 bit, target
- Makefile Adjusted for the new target
- Hardware UART addresses adjusted to the new target hardware
- boot.S and linker.ld, adjusted for 64 bit and align 8
- kernel.c/h syscall_handler fixed to evade pointer truncation
- Both HALs fixed to use the right addresses and offsets

+++ include/types.h
@@ -11,8 +11,8 @@ typedef signed short		i16_t;
 typedef signed int		i32_t;
 typedef signed long long	i64_t;
 
-typedef unsigned int		size_t;
-typedef unsigned int		uintptr_t;
-typedef signed int		intptr_t;
+typedef unsigned long long	size_t;
+typedef unsigned long long	uintptr_t;
+typedef signed long long	intptr_t;
 
 #endif
